requirements.txt
langchain
langchain_openai
google-generativeai
python-dotenv
opencv-python
pyaudio
google-cloud-speech
elevenlabs
SpeechRecognition
pydub
twilio